SB-138 (refactor) : User.class 패키지 이동
SB-138 (refactor) : User.class 패키지 이동
