+++ ai/src/main/java/ai/common/resolver/UserSessionResolver.java
@@ -2,7 +2,6 @@ package ai.common.resolver;
 
 import db.domain.user.enums.UserRole;
 import global.annotation.UserSession;
-import ai.domain.user.controller.model.User;
 import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
