+++ board/src/main/java/board/common/resolver/UserSessionResolver.java
@@ -1,6 +1,5 @@
 package board.common.resolver;
 
-import board.domain.user.controller.model.User;
 import db.domain.user.enums.UserRole;
 import global.annotation.UserSession;
 import java.util.Objects;
