+++ board/src/main/java/board/domain/board/controller/BoardApiController.java
@@ -2,24 +2,15 @@ package board.domain.board.controller;
 
 import board.common.response.MessageResponse;
 import board.domain.board.business.BoardBusiness;
-import board.domain.board.controller.model.detail.BoardDetailResponse;
 import board.domain.board.controller.model.register.BoardRegisterRequest;
 import board.domain.board.controller.model.register.BoardRegisterResponse;
-import board.domain.board.controller.model.search.BoardSearchResponse;
-import board.domain.board.controller.model.search.SearchCondition;
 import board.domain.board.controller.model.update.BoardUpdateRequest;
 import board.domain.board.controller.model.update.BoardUpdateResponse;
-import board.domain.user.controller.model.User;
+import board.common.resolver.User;
 import global.annotation.UserSession;
 import global.api.Api;
 import jakarta.validation.Valid;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
-import org.springframework.data.web.PageableDefault;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
